requirements.txt
psycopg2~=2.9.5
requests~=2.28.2
beautifulsoup4~=4.11.2
python-telegram-bot